models working properly, fixed maven dependencies for prersistence, added classes to persistence xml

+++ persistence/src/main/java/repositories/AuthorRepository.java
@@ -3,9 +3,8 @@ package repositories;
 import models.Author;
 
 import javax.persistence.EntityManager;
-import java.math.BigInteger;
 
-public class AuthorRepository extends GenericRepository<Author, BigInteger> {
+public class AuthorRepository extends GenericRepository<Author, Integer> {
 
     public AuthorRepository(EntityManager entityManager) {
         super(entityManager);
