+++ persistence/src/main/java/repositories/BookRepository.java
@@ -5,7 +5,7 @@ import models.Book;
 import javax.persistence.EntityManager;
 import java.math.BigInteger;
 
-public class BookRepository extends GenericRepository<Book, BigInteger> {
+public class BookRepository extends GenericRepository<Book, Integer> {
     protected BookRepository(EntityManager entityManager) {
         super(entityManager);
     }
